OCT-111: moved GetProductsQuery
