+++ components/catalogs/back/src/Application/Persistence/Catalog/Product/GetProductsQueryInterface.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Application\Persistence;
+namespace Akeneo\Catalogs\Application\Persistence\Catalog\Product;
 
 /**
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
