+++ components/catalogs/back/src/Infrastructure/Symfony/Resources/config/services.yml
@@ -128,7 +128,7 @@ services:
         arguments:
             - '@pim_catalog.repository.category'
 
-    Akeneo\Catalogs\Infrastructure\Persistence\GetProductsQuery:
+    Akeneo\Catalogs\Infrastructure\Persistence\Catalog\Product\GetProductsQuery:
         arguments:
             $getConnectorProducts: '@akeneo.pim.enrichment.product.connector.get_product_from_uuids'
